// src/front/src/api/identity/models/department.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Identity API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export type Department = (typeof Department)[keyof typeof Department];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Department = {
	Developers: "Developers",
	Support: "Support",
	Operations: "Operations",
	Finance: "Finance",
	Sales: "Sales",
	Others: "Others",
} as const;